cec-compliance: become temporarily a follower in the lost-msgs test
This test issues a lot of get-cec-version messages, and the remote
device will reply with cec-version messages. If we are not a follower,
and no other follower is running, then all these cec-version messages
are feature aborted by the CEC core, which messes up this test (too
much CEC traffic that can cause timeouts).

So become a follower, just for this test, so that all these replies
are just ignored.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>