cec-compliance: fix recognized/unrecognized_op logic
First of all cec_msg_status_is_abort() can't be used with CEC_RECEIVE,
instead check if the opcode is FEATURE_ABORT.

Secondly, don't assume that when a transmitted message times out, it is
not recognized. The reality is that the remote CEC device might be slow
or temporarily unresponsive (happens frequently when going in or out
of standby).

So only mark an op as (un)recoginized if it is unambiguous.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>