cec-ctl: the PA should be set before configuring the LAs
cec-ctl -p 0.0.0.0 --tv would set the LA to 14 if the physical
address before calling this command was != 0.0.0.0.

The problem was that the physical address was set after the adapter
was configured to a TV, and the CEC adapter will select LA 14 instead
of 0 if the physical address is != 0.0.0.0.

So first set the physical address, then claim the logical address(es).

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>