v4l2-ctl: don't use doioctl to query timings, use test_ioctl
The doioctl function will propagate any ioctl error as the application
result, but that's not what we want when querying timings/std. If that
fails we just continue on.

Also change the code so it looks at the v4l2_input capabilities to
decide whether this input support dv timings or std.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>