libdvbv5: add support for more PMT descriptors
There are several PMT descriptors that are now used. At least,
the HEVC (H-265) descriptor is now needed, in order to get the video
PID for some DVB-T2 providers in Europe.

So, sync most of those descriptors with the ones defined at the
VLC code. Please notice that, on VLC, it has descriptors also for
stream.type 0x06, 0x12, 0xa0 and 0xea, with are later detected,
by checking the descriptors inside the stream.

As we currently don't currently need to parse them, let's just
ignore for now.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>