v4l2-ctl: handle V4L2_EVENT_SOURCE_CHANGED
When the SOURCE_CHANGED event is received, go back to the beginning
of the capture loop, wait for a valid signal to appear and reallocate
buffers and continue streaming.

Note: this only works if --stream-poll is also set.
Otherwise the DQBUF will just block and you never see the
event.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>