v4l2-ctl: waiting for a valid timings only depends on no-query option
Right now we don't wait for valid timings when writing to a file or a
remote host. That makes no sense, since that is unrelated.

Also wait for valid timings first. Especially when writing to a remote
host it is important that the right timings/format information is set
and not changed later.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>