v4l2-ctl: add --disable-v4l2-ctl-stream-to configure option
The --stream-to and --stream-to-host options allow the user to
save the video to a file or stream it to a remote host. There are
situations where this is not something you want to support.

Add a new --disable-v4l2-ctl-stream-to configure option that disables
that --stream-to and --stream-to-host options. You can still test
streaming, just not save the video anywhere.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>