cec-compliance: try to reconnect if the HPD is gone
The transmit_timeout() function will fail if the HPD is gone.
This can happen when the remote TV switches inputs or goes into
standby/wakes up. Some TVs also pull the HPD low when in standby,
even though CEC still works. In that case you need to send an
Image View On message.

Modify the transmit_timeout() so that it waits for 3 seconds for
the HPD to reappear. If it doesn't, then it sends an Image View On
message and again waits for the HPD. If neither works, only then does
it fail.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>