ir-keytable: ensure udev rule fires on rc input device
The rc device is created before the input device, so if ir-keytable runs
too quickly the input device does not exist yet.

This also prevents udev from starting ir-keytable on an transmit only
device, which has no input device.

Note that $id in RUN will not work, since that is expanded after all the
rules are matched, at which point the the parent might have been changed
by another match in another rule. The argument to $env{key} is expanded
immediately.

Reported-by: Matthias Reichl <hias@horus.com>
Signed-off-by: Sean Young <sean@mess.org>