cec-ctl: fix topology tree view
The local adapter was no longer shows in the topology tree view that
is shown when the -S option is used.

In addition, when an adapter (local or remote) has multiple logical
addresses the logical address name was the same for both.

By encoding both PA and LA in the phys_addrs array the LA is preserved
and can be printed correctly.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>