cec-compliance: various improvements for no-HPD-in-standby
The standby tests failed for TVs that disable the HPD when in
standby. One reason was a silly mistake in a condition testing
for CEC_CAP_NEEDS_HPD: a ! was missing there.

But other issues were related to the tests themselves (don't
issue an IMAGE_VIEW_ON when trying to transmit a GIVE_POWER_STATUS,
STANDBY or ACTIVE_SOURCE command since that would defeat the purpose).

Another issue that the long_timeout handling was wrong. Use time()
to determine if we timed out, that's much more reliable compared to
the old method.

Tested with a display with this 'feature'.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>