v4l2-compliance: allow control ioctls without any controls
It is valid to support the control ioctls, but without any
controls. Make the control tests a bit less strict to allow for
this.

This situation can happen if a bridge driver makes a control
handler so subdevs can add their controls to it, but none of the
subdevs actually have controls.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Reported-by: Dave Stevenson <dave.stevenson@raspberrypi.org>