dvb_dev_get_fd(): return fd of local devices
This makes it possible to poll a local device.
Getting the fd is preferrable to adding a dvb_dev_poll() function,
because we can poll several fds together in an event-based program.

This is not implemented for remote devices, as polling a remote fd
does not make sense.
We could instead return the socket to know when to expect messages
from the remote device, but the current implementation in
dvb-dev-remote.c already runs a thread to receive remote messages
as soon as possible.

Signed-off-by: Rafaël Carré <funman@videolan.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>