dvb-sat: Use the right values for LNB setting
There's a bug at the logic under dvbsat_diseqc_set_input()
and dvb_sat_real_freq() with makes it to always use the
values for the universal LNBf when setting the input or
getting the real frequency.

The thing is that "lnb" is a very crap name for a static
array, and it shadows some local vars on other functions.

So, rename the array to "lnb_array" and be sure that those
two frequencies will read the lnb type from the params
struct.

This patch takes a suggestion made by Rafaël Carré,
expanding it to rename the lnb array, in order to avoid
further errors.

Reported-by: Rafaël Carré <funman@videolan.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>