dvb-sat: better handle out-of-range Satellite frequencies
There's no sense storing an entry if the frequency is out
of range. On satellite, this can easily be checked, as
the frequency should be compatible with the LNBf.

If such frequency is out of range, generate an alarm.

We should likely be also checking if the frequency matches
the one supported by the frontend, as informed by its
capabilities. For now, add it to our TODO list.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>