dvb: better cope with dvbloopback
The dvbloopback OOT Kernel module is a dirty hack. It requires
a Kernel patch that breaks the DVB core, removing
mutex_lock(&dvbdev_mutex) during DVB open.

Also, support for FE_GET_PROPERTY/FE_SET_PROPERTY is broken:
it causes the Kernel to crash with:

	dvblb_poll 0frontend1: fd:0
	BUG: unable to handle kernel paging request at 00007ffff41561c0
	IP: dvblb_ioctl+0x428/0x6e0 [dvbloopback]

Yet, it seems that users are trying to use it on some enviroments,
as I got some reports that libdvbv5 doesn't handle it well.

So, let's add some workarounds on libdvbv5, in order to make
it work.

There is a catch, though: without libdvbv5, it won't support
second generation TV standards, like DVB-S2 and DVB-T2.

This changeset works around 3 issues with dvbloopback:

1) only frontend0/dvr0/demux0/.. is used as a normal dvb adapter.
   The second ones are used by an special userspace interface,
   with shouldn't be vibible by libdvbv5 applications;
2) it expects that the device is opened with O_NONBLOCK;
3) the DVBv5 API support is broken. So, fallback to DVBv3.

Tested using https://github.com/bas-t/dvbloopback, with is
the first occurrence on Google for dvbloopback here.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>