cec/rds/v4l2-ctl: fix --list-devices option
All three utilities use the 'files.erase(iter)' construct inside a
for-loop over 'files'. This should be 'iter = files.erase(iter)'.

It works by accident only as long as you don't erase the last element,
after that iter is suddenly invalid. I discovered this with rds-ctl, but
the same problem is in all three utilities.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>