cec-ctl: fix bad end-of-bit handling
The code for --store-pin could insert an end-of-bit event with a timestamp
later than the next event after that, which messed up the analysis code.

Also improve the analysis code a bit to prevent it from unnecessarily adding
an end-of-bit event after reaching the end of the pin file.

It was harmless, but it looked weird.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>