dvbv5-zap: reset timeout at copy_to_file()
When a user selects a maximum time for recording a program,
he expects that the recorded program to be as close as
possible of the desired time. Right now, the time is accounted
since the beginning of tune, as it makes sense for other types
of timeout.

Make the recorded time closer to the desired timeout by
resetting the alarm counter when the first data is successfully
received.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>