dvbv5-zap: allow writing a file while monitoring the DVB stream
There's no good reason why it cannot be both monitoring and
writing into a file at the same time, provided that a valid
file name is provided.

Doing that helps to check what's going wrong with latency since
Kernel 4.9, where the ksoftirq code suffered a change.

So, add support for it.

Please notice that it can't accept stdout/stderr here, as dvbv5-zap
will be displaying the monitored data and errors at the console.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>