dvbv5-scan: print frontend stats on stdout on traffic monitor
While in traffic monitor, errors are directed to stderr, while
stats monitor to stdout. Yet, frontend stats were directed
to stderr as well. That makes harder to redirect stderr to
an error file, as it will print stats all the time.

So, change the logic to also print frontend stats on stdout.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>