v4l2-ctl/compliance: fix QUERYCTRL compat code for strings
If VIDIOC_QUERY_EXT_CTRL is not available, then these utilities
fall back to VIDIOC_QUERYCTRL and convert the result to a
v4l2_query_ext_ctrl struct.

But this means that the V4L2_CTRL_FLAG_HAS_PAYLOAD flag should be
set for strings. Unfortunately that was forgotten. This patch fixes
that.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>