v4l2-compliance: fix vivid test failure
The vivid radio devices can return different capabilities depending
on the rds_rx/tx_i_o_mode control values. After restoring the state
in v4l2-compliance we need to reopen the filehandle to get the
latest capabilities with QUERYCAP. Otherwise the read-write test
will fail because it thinks V4L2_CAP_READWRITE is not supported, when
in fact it is.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>