v4l2-compliance: add initial support for testing the MC
Add simple support for the media controller.

One could argue that this should go to a media-compliance test, but
that will just need to call v4l2-compliance in turn when you want to
test all the constituent devices in the topology. And since we currently
use it for 90% for V4L2, we might as well add it here.

Perhaps at some point v4l2-compliance will be renamed to media-compliance.

At least now we have a starting point that we can build on top of.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>