v4l2-compliance: add extra initial \n for error messages
When walking through the interfaces and we cannot (yet) test an interface,
then add an initial \n to the message so the text looks like this:

--------------------------------------------------------------------------------

Unable to detect what device /dev/dvb/adapter1/demux0 is, skipping.

--------------------------------------------------------------------------------

instead of this:

--------------------------------------------------------------------------------
Unable to detect what device /dev/dvb/adapter1/demux0 is, skipping.

--------------------------------------------------------------------------------

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>