v4l2-compliance: ignore colorspace tests for passthu subdevs
If the subdevice just passes through colorspace information from
source to sink, then do not test if this information is valid since
they will accept anything.

Note: this would be wrong for subdevs that do actual colorspace
conversion. We would have to test this in that case, but we probably
need properties first so we can create specific functions for this.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>