v4l2-dbg: improve usage messages
Drop the '=' separator between the option and the arguments in the
usage message. It's confusing and not needed.

Improve parsing of optional arguments: it can now understand
-l min=0x20,max=0x3f and --list-registers min=0x20,max=0x3f. This worked
fine for required arguments, but an '=' between the option and arguments
was needed if the arguments were optional, otherwise they wouldn't be
parsed.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>