cec-compliance: remove outdated comment
The comment about waiting X seconds before sending Image View On
was outdated. Just remove it. The other comment explaining why we
send Image View On was moved up a few lines to just before where
we create the CEC message instead of where we call the CEC_TRANSMIT
ioctl.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>