cec-compliance: if MAX_RETRIES and OK were set, clear OK
There was a bug in the CEC framework where tx_status could have
both OK and MAX_RETRIES flags set (this happened if a message
was canceled). The cec-compliance utility did warn about this, but
still passed tx_status on as-is. So clear the incorrect OK flag
instead to avoid problems with this.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>