cec-compliance: if MAX_RETRIES and OK were set, mark as RX timeout
Fix the previous fix. Upon closer examination of when both MAX_RETRIES
and OK are set it became clear that this happens when the TX succeeded,
but waiting for a reply was cancelled. So return an RX_TIMEOUT in that
case.

Note that this CEC framework bug was fixed in kernel 4.18.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>