configure.ac: disable gconv if it can't find its location
if --enable-gconv is used, it will need to know where gconv
files will be stored, otherwise linking the ARIB charset code
will fail.

Also, for it to work, v4l-utils should be built with shared
libraries support. there is already a check for it, except that
it doesn't really works the way it was coded.

Fix it by placing the check conditions at the right place and
adding a check to see if with_gconvdir was filled.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>