cec-compliance: fix set_osd_string test
Even though cec-follower implemented SET_OSD_STRING correctly,
cec-compliance would still give this warning:

warn: cec-test.cpp(332): The device is in an unsuitable state or cannot display the complete message.

This was due to an incorrect 'else' statement which should
have checked if the received message was a Feature Abort.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>