cec-compliance: some standby tests should be a warning
Tests were added to check the handling of core messages while
the remote device is in standby.

After testing with various devices I realized that many of those
do not handle this well, causing the compliance test to fail.

For now I have turned them into warnings. This might change in
the future. I feel that this falls in the category of "it is
really a failure, but almost everyone does this", so for the
time being a warning is more appropriate.

Signed-off-by: Hans Verkuil <hansverk@cisco.com>