cec-compliance: wake-up on Active Source is warn for <2.0
The CEC 2.0 specification explicitly forbids powering on a
display when Active Source is received. The older specification
does not explicitly forbid this (although it is really very clear
from the spec that you shouldn't do this).

Because the 1.4 spec was not crystal clear some devices expect a
display to power up when receiving Active Source, and so some
displays actually support this, thus failing this test.

So for 1.4 displays this test will report a warning, but it will
fail for 2.0 displays.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>