v4l2-ctl: fix set raw/sliced vbi format
Setting the raw or sliced VBI format didn't work correctly.
It was supposed to get the current format, then copy the
changed fields over to obtain the updated format. Instead
it used the format based on the values specified by the user,
i.e. it never used the updated current format.

As part of this patch the vbi_fmt(_out) variables were renamed
to sliced_fmt(_out) since it was not clear at all that vbi_fmt
referred to sliced VBI as opposed to raw VBI.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>