v4l2-compliance: add --stream-from and --stream-from-hdr options
One major missing piece of functionality has been to read the contents
of output buffers from a file. This made it impossible to test decoders
in particular.

New --stream-from and --stream-from-hdr (matching --stream-to and
--stream-to-hdr from v4l2-ctl) options have been added to support this.

It is also possible to specify files for specific pixelformats by
prefixing the filename with '<fourcc>=', i.e.:

FWHT=test.fwht

Multiple files for different pixelformats can be given this way.

The file given without a prefix is used as the default file if no
file for the current pixelformat was found.

Note: the compliance test for codecs is still under development and
it may not always work.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>