qvidcap: split up capture-win-gl.cpp into capture.cpp and paint.cpp
The capture-win-gl.cpp was too large, so it is now split in one
part (capture.cpp) that deals with streaming and one (paint.cpp)
that deals with the openGL shader and render code.

It has also been renamed since the -win-gl bit was really a
leftover from qv4l2 and made no sense for the openGL-only qvidcap.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>