v4l2-compliance: work around vivid error injection controls
v4l2-compliance refuses to work with vivid if vivid is loaded without
the no_error_inj=1 module option since some of the compliance tests
trigger the error injection, thus failing the test.

Make v4l2-compliance smarter by ignoring all vivid class controls if
it detects that the driver is in fact vivid. This allows it to
successfully pass all tests.

The main reason for this change is that this allows us to use the
error injection controls to test how the kernel handles various error
conditions that are otherwise hard to test.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>