cec-compliance: fix wrong test for no HPD in testRemote()
If the 'standby' flag for the test-to-be-performed is set to true,
then the code needs to check is the remote device pulls down the
HPD when in standby. In that case the test should be skipped since
it is not a valid test for such devices.

However, the check was wrong and these tests were still performed.

Fix the check.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>