v4l2-compliance: better field and prepbuf/qbuf tests
When queueing a buffer to an OUTPUT device with V4L2_FIELD_ANY, then
VIDIOC_QBUF shall return a non-ANY field value. Check this.

If a buffer has been prepared before VIDIOC_QBUF was called, then
for USERPTR and DMABUF the userpointers and dmabuf fds are ignored
and instead it will return the ones used when VIDIOC_PREPARE_BUF was
called. Check this.

Signed-off-by: Hans Verkuil <hansverk@cisco.com>