cec-ctl: fix Remote Control Passthrough Feature
This feature didn't show up in the --help due to a parsing bug.
This feature is unusual because it only has 'Has also' messages
and that was something the parses couldn't handle.

Also add an extra newline when printing the feature name in the
help to make it easier to read.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>