media-ctl: support a bus-info string as argument to -d
If the device passed to the -d option is not found, then interpret it
as a bus-info string and try to open all media devices and see which one
reports a bus-info string equal to the -d argument.

That makes it possible to open a specific media device without having to know
the name of the media device.

Similar functionality has been implemented for v4l2-ctl and v4l2-compliance,
and for the cec utilities.

This allows scripts that no longer need to care about the name of a device
node, instead they can find it based on a unique string.

Also extend the -d option to support -d0 as a shorthand for /dev/media0 to
make it consistent with the other utils.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>