test-media: replace rmmod by separate unbind and rmmod
If kobjects are delayed-released, then when using rmmod the
module memory can be freed before the kobject release function
is called, which might be in that freed memory.

As far as I can tell there is no protection for this, and so
errors due to this are not a bug in the driver, but a consequence
of enabling this debug option.

So use unbind to check for issues with delayed release. And
run rmmod once all those objects are safely released.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>