test-media: add sleeps after configuring CEC
Add a sleep 1 after configuring each vivid CEC adapter to give it
time to settle.

Without sleeps this can hit a limitation of the vivid CEC emulation
where transmits can repeatedly fail with the ARB_LOST error and
as a result it picks Playback Device 2 instead of 1, causing the
remainder of the CEC tests to fail.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>