v4l2-compliance: fix check for entity names
sizeof(name) is 64 for media_v2_entity but 32 for
struct media_entity_desc. So only consider the first 32 bytes
(31 characters) for the duplicate name tests and matching what
ENUM_ENTITIES reports with what G_TOPOLOGY reports.

Also relax the check of the possible interface types to allow
alsa as this is upcoming with the Media Device Allocator API.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>