dvbv5-zap: improve program exit code
While testing the program with Valgrind, it would keep reporting
about non-freed data. The main reason is that, when called in
non-record and non-monitor mode, the while() loop doesn't check
for the timeout indication.

Solve that and be sure that all strdup() vars will be freed
too.

On my tests, on normal mode, on monitor mode and on record mode,
it now shows no memory leaks with Valgrind.

Having zero leaks in Valgrind is important mostly because we
don't want the libraries to leak. Having all memory allocated
internally at the program freed means that, if Valgrind will
report any memory leaks in the future, it would belong to
the library code.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>