cec-ctl: fix --stress-test-power-cycle
This test transmitted Image View On from LA 0xf. Which is fine if
this is a display that pulls down the HPD when in Standby, but it
fails for all other displays. In that case you just use the current
logical address.

Fix this by first trying to transmit with the current LA, then on
error transmit again, now from LA 0xf.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>