v4l2-ctl: improve stream-count handling for m2m
When streaming for m2m devices the stream-count should be exact
and work even for e.g. --stream-count=1.

This means that when filling in the initial output buffers, no more
than stream_count buffers should be filled, even for count values
less than the number of allocated buffers.

This also means that STOP should be issued right after filling in
the buffers if stream_count has already reached 0.

This patch also is more precise when > or < is printed and when
the buffer contents (--verbose) is shown.

'<' is shown after a capture buffer is dequeued.
'>' is shown after an output buffer is queued.

The buffer contents is always shown after a buffer is dequeued.

For capture or output devices the stream count refers to the number
of dequeued buffers, and the number of queued buffers is
more than that (number of allocated buffers + stream_count).

For m2m devices the stream_count refers to the number of queued
output buffers.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>