v4l2-ctl: fixes relating to frame counting and draining
do_handle_cap assumed that it could skip frame counting for m2m
devices, but that's not true. It should count frames for
stateful encoders and for the cap2out streaming use-case.

So make this an argument to do_handle_cap instead.

Also fix various issues with detecting when the output side of
an m2m device is stopped (e.g. because --stream-count buffers
have been queued up).

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>