v4l2-ctl: disable_trace disabled too much
When trying to stream twice from the same device the second attempt
silently exits since REQBUFS returns EBUSY. The problem is that there
is no feedback of the ioctl error to the user, and that is because
the ioctl tracing is disabled before REQBUFS is called.

Change the code to be a bit more refined w.r.t. what is traced and
what isn't.

Now a proper error code is shown.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>