v4l2-compliance: if G_FMT returns 0, mark buftype as valid
Currently a buftype was only marked as valid in node->valid_buftypes
if G_FMT returned 0 AND the validation of the returned v4l2_format
was also successful. But this gave bad knock-on effects in later tests.

So mark the buftype as valid if G_FMT returns 0, independent of the
later tests.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>