keytable: load bpf protocol for unsupported protocols
If there is no hardware decoder or kernel software decoder available
for a protocol, attempt to load a bpf protocol for this.

This opens the door for building the kernel without IR decoders (non
bpf ones) and just rely on bpf and hardware decoders. This requires
the bpf decoders to be written though, this still needs to be done.

Signed-off-by: Sean Young <sean@mess.org>