v4l-stream.c: add missing copy_cap_to_ref() and fix state
The fwht decoder part was missing the copy_cap_to_ref(): this copies
the decoded frame back to the internal reference frame. This used to
be done in the fwht decoder itself in the past, but after adding
stateless decoder support to vicodec this step was split off into
a separate function.

Also update the state structure initialization: it didn't set
ref_stride and was missing support for 1 or 4 components.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>