v4l2-compliance: add timeout when waiting for event
The test for the stateful decoder that checks for a SOURCE_CHANGE
event when you start streaming needs to wait for up to one second
before calling dqevent. Otherwise the HW may not have had the
chance yet to raise this event.

So use select() to wait for an exception with a 1s timeout.

Reported-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>