cec-ctl: add --test-power-cycle option
Add the new --test-power-cycle option: this will test how well a
display supports the Standby and Image View On messages. This is
the first version, and it is likely to be improved going forward.

Also improved the --stress-test-power-cycle option: it is now
possible to provide a max-sleep suboption to have the test sleep
for a random time from 0 to max-sleep seconds before transmitting
Standby or Image View On.

This will randomize the timings.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>