v4l2-compliance: fix m2m field check
The 'field' value of a buffer is checked against cur_fmt, but for
m2m devices there are two different formats: one for capture and
one for output. So for m2m capture devices check against cur_m2m_fmt
instead.

During streaming another issue occurs when it checks the 'field' of
the capture buffer against the field of the original output buffer,
but these can differ for e.g. m2m deinterlacers. So relax that test.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>