v4l2-compliance: fix USERPTR test if PREPARE_BUF is not supported
If the driver under test does not support VIDIOC_PREPARE_BUF, then
the USERPTR streaming test will fail since length of the buffer was
zeroed. But that should only be done if PREPARE_BUF was supported,
so move the code that zeroes length, bytesused and userptr under
the 'if (ret == 0)' (the VIDIOC_PREPARE_BUF return code).

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>