keytable: add support for keymap with raw literals
These rc keymaps look like:

	[[protocols]]
	name = 'Beon_Colour_TV'
	protocol = 'raw'
	[[protocols.raw]]
	keycode = 'KEY_POWER'
	raw = [ 1324, 842, 6697, 26400, 1324, 842, 6697 ]
	[[protocols.raw]]
	keycode = 'KEY_MUTE'
	raw = [ 1305, 865, 2357, 931, 3411, 26400, 1305, 865, 2357, 931, 3411 ]

So the incoming IR will be literally compared against the IR given, and only
if the entire array matches with no leading or trailing IR will the keycode
be reported.

This is analogous to lircd raw_codes support, so lircd2toml.py can also
convert raw_codes lircd.conf files. This means that the vast majority of
lircd files can now be converted.

This feature requires loops in BPF programs so this is only supported
in kernel 5.3 onwards.

Signed-off-by: Sean Young <sean@mess.org>