v4l2-compliance: split node at testNode() in two nodes
v4l2-compliance tests deals with memory-to-memory devices (m2m) like
they are a single device, through the struct node. This is indeed the
most usual case. However, there are m2m devices that have two
separated devices for the capture queue and for the output queue, and
the current design does not support they.

Split the node on testNode and related functions in two arguments:
struct node node and struct node node_m2m_cap. For now, use the same node in
both arguments.

Signed-off-by: André Almeida <andrealmeid@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>