ir-ctl: support raw format by default
When "ir-ctl -r" outputs data it prints one line per space and pulse. This
is not very space efficient and long protocols like rc-6 are unlikely to
fit on the screen.

The new format is much more compact. A single IR message is printed on a
single line. The pulse is prefixed with an optional + and a space is
prefixed with -.

The rc-5 example from the man page becomes:

$ ir-ctl -r
+889 -889 +1778 -1778 +889 -889 +889 -889 +889 -889 +1778 -889 +889 -889 +889 -889 +889 -889 +889 -889 +889 -1778 +889 # timeout 133889

The old behaviour can be resurrected with: ir-ctl -r --mode2

This commit also add support for sending the same format.

Signed-off-by: Sean Young <sean@mess.org>