keytable: ensure we have enough memlock pages
Since kernel v5.2, BPF maps and programs are charged against
RLIMT_MEMLOCK. By default this limit is 64KB however all of these are
already taken (16 pages on x86) by the time we've booted on Fedora. This
results in a permission denied error.

The error message is confusing since it happens when running as root.
systemd works around this problem by setting setrlimit(RLIMIT_MEMLOCK)
to 64MB. Do the the same.

Signed-off-by: Sean Young <sean@mess.org>