cec: fix cec-log.h dependency issues
Due to a reorg, cec-log.h has now a different meaning and it
is placed on a different directory: utils/common/cec-log.h.

The problem is that it now conflicts with the old auto-generated
one at the utils/cec-follower directory.

This won't affect new clones, but if someone updates v4l-utils,
the build will fail without explicitly removing the old file.

So, let's change the includes for them to explicitly pinpoint
to the right cec-log.h file. This way, if the old one exists,
it won't cause any build breakages.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>