cec-follower: rename freq_idx to service_idx
The implementation of digital support in tuner control requires the
renaming of freq_idx to service_idx as analog and digital channels
are referred to collectively as services.  The index will now represent
the current state of both analog and digital.  Tuner step features will
cycle through both analog and digital with digital preceeding analog.

Signed-off-by: Jiunn Chang <c0d1n61at3@gmail.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>