v4l2-compliance: fix read/write tests
The read test failed for metadata capture if that is not supported
by the current input (read returns EINVAL in that case).

Allow for such situations.

The same situation occurs for VBI, but that worked since there
was a specific check for VBI. Improve the test for VBI as well, verifying
that read()/write() for a VBI device fails if no VBI is supported for the
current input/output.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>