cec-compliance: add -e and -l options
The new -l (--list-tests) option lists all tests that are run,
and a list of possible test results.

The new -e (--expect) option allows the user to specify expected
test results. This makes it possible to e.g. check that an optional
feature is actually supported, or that a FAIL result is actually
expected.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>